Updating CPU usages

+++ WindowsPerformanceMonitor/Models/ProcessEntry.cs
@@ -7,7 +7,7 @@ namespace WindowsPerformanceMonitor.Models
     public class ProcessEntry
     {
         public string Name { get; set; }
-        public double Pid { get; set; }
+        public int Pid { get; set; }
         public double Cpu { get; set; }
         public double Gpu { get; set; }
         public double Memory { get; set; }
